Unique Years
Counting the number of years (or streak) a donor has given over a period of time is not built into Salesforce and its actually really difficult to achieve - conceptually and with SOQL Limit errors. This project was to calculate the number of years a donor has given over a period of time.